Some renamings of tsolver example
